Add content for demarches and a concrete example with Glasgow
Start of the info for demarches, missing at least the languages tests, and statements of purpose.
I need to add the list of current agreements, plus the list of university where for one reason or another we can't go. (Only Edinburgh that I know of)
I did a concrete example with the University of Glasgow, but since I haven't been there yet I can't fill everything. But with this and the template, people should get the gist of it.