Add further information regarding the language tests as well as the list of partner universities.
Missing the Statement of purpose, not sure I'll do it or if it's actually useful.
I need to correct the paths to work when the site is not in the server root directory, silly of me.
Need proofreading and quality check.
Annnnnd, now I must do a small page for every university. Yay ! I hope someone will help me.
I'm afraid it won't happen though. Just like no one will ever read this. If you do, I'll offer you a beer. Yay ! (If you actually contribute to the site, that is)